combine deserialize modules
