+++ src/lib/util/deserialize.ts
@@ -1,14 +1,21 @@
 import {traverse} from '@feltcoop/felt/util/object.js';
 
-import type {Deserializers} from '$lib/util/deserializers';
+// TODO where does this belong? vocab utils?
 
-// TODO BLOCK where does this belong? vocab utils?
+export const deserializeDate = (v: string): Date => new Date(v);
+
+export type Deserializers = Map<string, (v: any) => any>;
+
+export const deserializers: Deserializers = new Map([
+	['created', deserializeDate],
+	['updated', deserializeDate],
+]);
 
 export interface DeserializeProperties {
 	(value: any): void;
 }
 
-export const deserializeProperties =
+export const deserialize =
 	(deserializers: Deserializers): DeserializeProperties =>
 	(value) => {
 		traverse(value, (key, value, obj) => {
