+++ src/lib/util/testAppHelpers.ts
@@ -8,8 +8,7 @@ import {toDispatch} from '$lib/app/dispatch';
 import {findHttpService} from '$lib/ui/services';
 import {installSourceMaps} from '$lib/util/testHelpers';
 import {mutations} from '$lib/app/mutations';
-import {deserializeProperties} from '$lib/util/deserializeProperties';
-import {deserializers} from '$lib/util/deserializers';
+import {deserialize, deserializers} from '$lib/util/deserialize';
 
 installSourceMaps();
 
@@ -24,7 +23,7 @@ export const setupApp =
 		const ui = toUi(session, false, {});
 		const httpApiClient = toHttpApiClient<EventParamsByName, EventResponseByName>(
 			findHttpService,
-			deserializeProperties(deserializers),
+			deserialize(deserializers),
 			fetch,
 		);
 		context.app = {
